upnp.py: rename and prepare for other protocols
